pyproject.toml
[build-system]
# A list of packages that are needed to build your package:
requires = ["setuptools"] # REQUIRED if [build-system] table is used
# The name of the Python object that frontends will use to perform the build:
build-backend = "setuptools.build_meta" # If not defined, then legacy behavior can happen.


[project]
name = "arxiv_explorer"
version = "0.0.1" # REQUIRED, although can be dynamic
description = "A vector-lookup based Arxiv search engine."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE.txt" }
keywords = ["sample", "setuptools", "development"]
authors = [{ name = "Tom Carter", email = "tomcarter017@gmail.com" }]
maintainers = [
  { name = "Tom Carter", email = "tomcarter017@gmail.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = ["huggingface_hub", "sentence-transformers", "faiss-cpu", "numpy", "pymongo", "tqdm"]

[project.optional-dependencies]
test = ["pytest"]
data = ["kagglehub"]

[project.scripts]
sample = "arxiv_explorer:main"

